pyproject.toml
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["nelo_data_engg"]
known_third_party = ["boto3", "pandas", "duckdb", "yaml", "dotenv"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "botocore.*",
    "duckdb.*",
    "pandas.*",
    "yaml.*",
]
ignore_missing_imports = true

[project]
name = "nelo-data-engg"
version = "0.1.0"
description = "Nelo Growth Analytics Data Pipeline"
requires-python = ">=3.9"
dependencies = [
    "boto3==1.40.51",
    "python-dotenv==1.1.1",
    "duckdb==1.4.1",
    "pandas==2.3.3",
    "PyYAML==6.0",
    "diskcache==5.6.3",
    "pre-commit==3.6.0",
    "black==24.1.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "shellcheck-py==0.10.0.1",
    "sqlfluff==2.3.5",
    "great-expectations==0.18.8",
    "pytest==8.4.2",
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.0",
    "black==24.1.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "shellcheck-py==0.10.0.1",
    "sqlfluff==2.3.5",
    "great-expectations==0.18.8",
    "pytest==8.4.2",
]

[tool.hatch.build.targets.wheel]
packages = ["."]